Reduce internal line materialization churn
