+++ benchmarks/safari.json
@@ -3,7 +3,7 @@
   "results": [
     {
       "label": "Our library: prepare()",
-      "ms": 20.500000000000057,
+      "ms": 18.5,
       "desc": "One cold 500-text measurement batch"
     },
     {
@@ -13,12 +13,12 @@
     },
     {
       "label": "DOM batch",
-      "ms": 88.5,
+      "ms": 88.00000000000011,
       "desc": "Single 400→300px batch resize: write all, then read all"
     },
     {
       "label": "DOM interleaved",
-      "ms": 154,
+      "ms": 152.00000000000045,
       "desc": "Single 400→300px batch resize: write + read per div"
     }
   ],
@@ -40,14 +40,14 @@
     },
     {
       "label": "Our library: layoutNextLine()",
-      "ms": 0.05,
+      "ms": 0.025000000000022737,
       "desc": "60-text shared-corpus batch across widths 180/220/260px; streaming line-by-line layout"
     }
   ],
   "richInlineResults": [
     {
       "label": "Our library: measureRichInlineStats()",
-      "ms": 0.024999999999988632,
+      "ms": 0.025,
       "desc": "36 rich-text inline flow shared-corpus texts across widths 180/220/260px; stats only"
     },
     {
@@ -62,14 +62,14 @@
     },
     {
       "label": "Our library: layoutNextRichInlineLineRange() + materializeRichInlineLineRange()",
-      "ms": 0.05,
+      "ms": 0.049999999999977264,
       "desc": "36 rich-text inline flow shared-corpus texts across widths 180/220/260px; streaming range walk plus per-line materialization"
     }
   ],
   "richPreWrapResults": [
     {
       "label": "Our library: measureLineStats()",
-      "ms": 0.25000000000002276,
+      "ms": 0.25,
       "desc": "12 generated pre-wrap texts with 320 hard-break chunks across widths 220/260/320px; stats only"
     },
     {
@@ -84,29 +84,29 @@
     },
     {
       "label": "Our library: layoutNextLine()",
-      "ms": 0.95,
+      "ms": 1.050000000000023,
       "desc": "12 generated pre-wrap texts with 320 hard-break chunks across widths 220/260/320px; streaming line-by-line layout"
     }
   ],
   "richLongResults": [
     {
       "label": "Our library: measureLineStats()",
-      "ms": 1.4499999999999886,
+      "ms": 1.45,
       "desc": "8 Arabic long-form texts across widths 240/300/360px; stats only"
     },
     {
       "label": "Our library: layoutWithLines()",
-      "ms": 4.724999999999989,
+      "ms": 4.8,
       "desc": "8 Arabic long-form texts across widths 240/300/360px; materializes text lines"
     },
     {
       "label": "Our library: walkLineRanges()",
-      "ms": 2.4,
+      "ms": 2.3499999999999774,
       "desc": "8 Arabic long-form texts across widths 240/300/360px; line ranges only, no line text strings"
     },
     {
       "label": "Our library: layoutNextLine()",
-      "ms": 4.325,
+      "ms": 4.375000000000045,
       "desc": "8 Arabic long-form texts across widths 240/300/360px; streaming line-by-line layout"
     }
   ],
@@ -122,8 +122,8 @@
       "width": 300,
       "lineCount": 193,
       "analysisMs": 2,
-      "measureMs": 5,
-      "prepareMs": 7,
+      "measureMs": 4,
+      "prepareMs": 6,
       "layoutMs": 0.02
     },
     {
@@ -137,7 +137,7 @@
       "width": 300,
       "lineCount": 380,
       "analysisMs": 5,
-      "measureMs": 6.999999999996362,
+      "measureMs": 6,
       "prepareMs": 11,
       "layoutMs": 0.04
     },
@@ -151,8 +151,8 @@
       "breakableSegments": 0,
       "width": 300,
       "lineCount": 428,
-      "analysisMs": 3,
-      "measureMs": 9,
+      "analysisMs": 4,
+      "measureMs": 8,
       "prepareMs": 12,
       "layoutMs": 0.08
     },
@@ -166,9 +166,9 @@
       "breakableSegments": 2,
       "width": 300,
       "lineCount": 332,
-      "analysisMs": 3,
-      "measureMs": 9,
-      "prepareMs": 11,
+      "analysisMs": 2,
+      "measureMs": 8.000000000001819,
+      "prepareMs": 10.000000000001819,
       "layoutMs": 0.06
     },
     {
@@ -182,8 +182,8 @@
       "width": 300,
       "lineCount": 626,
       "analysisMs": 6.999999999998181,
-      "measureMs": 14,
-      "prepareMs": 20,
+      "measureMs": 12,
+      "prepareMs": 18.00000000000182,
       "layoutMs": 0.06
     },
     {
@@ -197,8 +197,8 @@
       "width": 300,
       "lineCount": 375,
       "analysisMs": 4,
-      "measureMs": 7,
-      "prepareMs": 12,
+      "measureMs": 7.999999999996362,
+      "prepareMs": 11.000000000001819,
       "layoutMs": 0.035
     },
     {
@@ -211,9 +211,9 @@
       "breakableSegments": 8087,
       "width": 300,
       "lineCount": 1024,
-      "analysisMs": 8.999999999998181,
-      "measureMs": 24.00000000000182,
-      "prepareMs": 33,
+      "analysisMs": 8,
+      "measureMs": 23.00000000000182,
+      "prepareMs": 32,
       "layoutMs": 0.095
     },
     {
@@ -227,8 +227,8 @@
       "width": 300,
       "lineCount": 81,
       "analysisMs": 1,
-      "measureMs": 4,
-      "prepareMs": 5,
+      "measureMs": 3.000000000001819,
+      "prepareMs": 4.000000000001819,
       "layoutMs": 0.005
     },
     {
@@ -243,7 +243,7 @@
       "lineCount": 54,
       "analysisMs": 0,
       "measureMs": 3,
-      "prepareMs": 3,
+      "prepareMs": 3.000000000001819,
       "layoutMs": 0.005
     },
     {
@@ -272,8 +272,8 @@
       "width": 300,
       "lineCount": 591,
       "analysisMs": 8,
-      "measureMs": 13.000000000001819,
-      "prepareMs": 22,
+      "measureMs": 14.999999999996362,
+      "prepareMs": 22.00000000000182,
       "layoutMs": 0.085
     },
     {
@@ -288,7 +288,7 @@
       "lineCount": 653,
       "analysisMs": 5.000000000001819,
       "measureMs": 30.00000000000182,
-      "prepareMs": 36.99999999999818,
+      "prepareMs": 35.99999999999818,
       "layoutMs": 0.075
     },
     {
@@ -302,9 +302,9 @@
       "width": 300,
       "lineCount": 2860,
       "analysisMs": 9,
-      "measureMs": 105.99999999999818,
-      "prepareMs": 114.99999999999818,
-      "layoutMs": 0.455
+      "measureMs": 103.00000000000182,
+      "prepareMs": 112,
+      "layoutMs": 0.44
     },
     {
       "id": "ar-risalat-al-ghufran-part-1",
@@ -316,11 +316,11 @@
       "breakableSegments": 18745,
       "width": 300,
       "lineCount": 2644,
-      "analysisMs": 19,
+      "analysisMs": 18,
       "measureMs": 180,
-      "prepareMs": 198,
+      "prepareMs": 199,
       "layoutMs": 0.28
     }
   ],
-  "requestId": "1775740912323-fzmlunetfci"
+  "requestId": "1775742828351-yp1l141vul8"
 }
